Preliminary implementation of sleep/wakeup.

+++ src/sys/kerninc/InvParam.h
@@ -102,7 +102,10 @@ get_iparam32(InvParam_t* iParam)
  * pased in registers is passed in an even-numbered parameter word pair.
  * Regardless of native calling convention, the lower-numbered
  * parameter word contains the portion of the value at the least
- * significant address */
+ * significant address.
+ *
+ * If this implementation changes, be sure to update the rewriting
+ * logic in the kernel sleep capability implementation. */
 static inline uint64_t
 get_iparam64(InvParam_t* iParam)
 {
@@ -113,7 +116,7 @@ get_iparam64(InvParam_t* iParam)
     return get_pw(iParam->invoker, iParam->next_idw-1);
   }
   else if (sizeof(uintptr_t) == 4) {
-    if ( (alignof(uintptr_t) == 8) && ((iParam->next_idw % 2) != 0) )
+    if ( (alignof(uint64_t) == 8) && ((iParam->next_idw % 2) != 0) )
       iParam->next_idw++;
 
     iParam->next_idw += 2;
@@ -154,7 +157,7 @@ put_oparam64(InvParam_t* iParam, uint64_t v)
     iParam->opw[iParam->next_odw++] = v;
   }
   else if (sizeof(uintptr_t) == 4) {
-    if ( (alignof(uintptr_t) == 8) && ((iParam->next_odw % 2) == 1) )
+    if ( (alignof(uint64_t) == 8) && ((iParam->next_odw % 2) == 1) )
       iParam->next_odw++;
 
     iParam->opw[iParam->next_odw++] = ((uint32_t *) &v)[0];
